feat(auth/password-reset): jwt password reset token strategy and payload
- Moving the strategies and payloads to application layer
