+++ libs/core/security/src/lib/authentication/application/index.ts
@@ -3,4 +3,5 @@ export * from './services';
 export * from './commands';
 export * from './guards';
 export * from './validations';
+export * from './strategies';
 export * from './dtos';
