+++ libs/core/security/src/lib/authentication/application/jwt-payloads/jwt-password-reset.payload.ts
@@ -1,0 +1,7 @@
+export interface JwtPasswordResetPayload {
+    email: string;
+    issuedAt: number;
+    expiration: number;
+}
+
+export type JwtPasswordResetPayloadSign = Omit<JwtPasswordResetPayload, "issuedAt" | "expiration">;
