+++ libs/core/security/src/lib/authentication/application/strategies/index.ts
@@ -1,0 +1,2 @@
+export * from './jwt-auth.strategy';
+export * from './password-reset.strategy';
