--- libs/core/security/src/lib/authentication/domain/strategy/index.ts
@@ -1,1 +1,0 @@
-export * from './jwt.strategy';
